% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_wt.R
\name{read_geoda}
\alias{read_geoda}
\title{Read A Geoda File(.gal,.gwt,.swm)}
\usage{
read_geoda(file_path, id_vec = NULL)
}
\arguments{
\item{file_path}{The file paht of the geoda file.}

\item{id_vec}{(optional),the id_vec is the id values used in the geoda file.}
}
\value{
A weights object
}
\description{
Create a spatial weights object from a geoda file
}
\author{
Wenbo Lv
}
