% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrgeoda_helper.R
\name{scale_fill_lisa}
\alias{scale_fill_lisa}
\title{LISA Fill Scales}
\usage{
scale_fill_lisa(name = "LISA", ...)
}
\arguments{
\item{name}{The name of the LISA fill scales legend,default is \code{LISA}}

\item{...}{Adjust other legend details for the LISA fill scales, like \code{labels}
to adjust the appearance of legend, see \code{?ggplot2::scale_fill_manual()}.}
}
\description{
Provide \code{ggplot2} fill scales like geoda software.Now it achieve by using
\code{?ggplot2::scale_fill_manual()}.Another achieve can see
https://stackoverflow.com/questions/43440068/ggplot2-fix-colors-to-factor-l.
}
\examples{
library(sf)
library(ggplot2)
guerry_path = system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry = read_sf(guerry_path)
guerry |>
 dplyr::mutate(lisa = st_local_moran(guerry,'Crm_prs')) |>
 dplyr::select(lisa) |>
 ggplot() +
 geom_sf(aes(fill = lisa),lwd = .1,color = 'grey') +
 scale_fill_lisa()
}
\author{
Wenbo Lv
}
