% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{st_azp_sa}
\alias{st_azp_sa}
\title{A simulated annealing algorithm to solve the AZP problem}
\usage{
st_azp_sa(
  sfj,
  varcol,
  k,
  wt = NULL,
  boundvar = NULL,
  cooling_rate = 0.85,
  sa_maxit = 1,
  min_bound = 0,
  inits = 0,
  initial_regions = vector("numeric"),
  scale_method = "standardize",
  distance_method = "euclidean",
  seed = 123456789,
  rdist = numeric()
)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{varcol}{The variable selected to calculate spatial lag, which is a character.}

\item{k}{The number of clusters.}

\item{wt}{(optional) The spatial weights object,which can use \code{st_weights()} to
construct,default is constructed by \code{st_weights(sfj,'contiguity')}.}

\item{boundvar}{(optional) A data frame / tibble with selected bound variable.}

\item{cooling_rate}{(optional) The cooling rate of a simulated annealing algorithm. Defaults to 0.85.}

\item{sa_maxit}{(optional) The number of iterations of simulated annealing. Defaults to 1.}

\item{min_bound}{(optional) A minimum bound value that applies to all clusters.}

\item{inits}{(optional) The number of construction re-runs, which is for ARiSeL
"automatic regionalization with initial seed location".}

\item{initial_regions}{(optional) The initial regions that the local search starts with.
Default is empty. means the local search starts with a random process to "grow" clusters.}

\item{scale_method}{(optional) One of the scaling methods 'raw', 'standardize',
'demean', 'mad', 'range_standardize', 'range_adjust' to apply on input data.
Default is 'standardize' (Z-score normalization).}

\item{distance_method}{(optional) The distance method used to compute the distance betwen
observation i and j. Defaults to "euclidean". Options are "euclidean" and "manhattan"}

\item{seed}{(optional) The seed for random number generator. Defaults to 123456789.}

\item{rdist}{(optional) The distance matrix (lower triangular matrix, column wise storage).}
}
\value{
A names list with names "Clusters", "Total sum of squares", "Within-cluster sum of squares",
"Total within-cluster sum of squares", and "The ratio of between to total sum of squares".
}
\description{
A wrapper function for \code{rgeoda::azp_sa()}.The automatic zoning procedure (AZP) was
initially outlined in Openshaw (1977) as a way to address some of the consequences of
the modifiable areal unit problem (MAUP). In essence, it consists of a heuristic to
find the best set of combinations of contiguous spatial units into p regions, minimizing
the within sum of squares as a criterion of homogeneity. The number of regions needs to
be specified beforehand.
}
\examples{
library(sf)
guerry = read_sf(system.file("extdata", "Guerry.shp", package = "rgeoda"))
guerry_clusters = st_azp_sa(guerry,c('Crm_prs','Crm_prp','Litercy','Donatns',
'Infants','Suicids'),5)
guerry_clusters
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
