% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrgeoda_helper.R
\name{st_breaks}
\alias{st_breaks}
\title{Univariate Spatial Stratification}
\usage{
st_breaks(sfj, varcol, break_method = "stddev", k = 6)
}
\arguments{
\item{sfj}{An sf, tibble or data.frame object}

\item{varcol}{The variables selected to run univariate spatial stratification.}

\item{break_method}{(optional) Which has to be one of "stddev"(default), "hinge15",
"hinge30", "percentile", "natural", "quantile". When the \code{break_method} is "natural"
or "quantile", you can assign the \code{k} argument to have how many breaks, other methods
will have 6 groups.}

\item{k}{(optional)A numeric value indicates how many breaks,default is 6.}
}
\value{
A vector of numeric values of computed breaks
}
\description{
Univariate Spatial Stratification by invoking rgeoda's *_breaks function.
}
\examples{
library(sf)
guerry = read_sf(system.file("extdata", "Guerry.shp", package = "rgeoda"))
st_breaks(guerry,'Crm_prs',break_method = "quantile", k = 5)
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
