% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_wt.R
\name{st_contiguity_weights}
\alias{st_contiguity_weights}
\title{Contiguity Spatial Weights}
\usage{
st_contiguity_weights(
  sfj,
  queen = TRUE,
  order = 1,
  include_lower_order = FALSE,
  precision_threshold = 0
)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{queen}{(Optional) TRUE (default) or FALSE, TRUE implements Queen
Contiguity and FALSE implements Rook Contiguity.}

\item{order}{(Optional) Order of contiguity, default is 1.}

\item{include_lower_order}{(Optional)  Whether or not the lower order
neighbors should be included in the weights structure,default is False.}

\item{precision_threshold}{(Optional) The precision of the underlying shape
file is insufficient to allow for an exact match of coordinates to determine
which polygons are neighbors,default is 0.}
}
\value{
An instance of rgeoda Weight-class.
}
\description{
Create a contiguity spatial weights with options of "queen", "order",
"include lower order" and "precision threshold"
}
\examples{
guerry_path = system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry = sf::read_sf(guerry_path)
queenw = st_contiguity_weights(guerry,queen = TRUE)
}
\author{
Wenbo Lv
}
