% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_wt.R
\name{st_distance_weights}
\alias{st_distance_weights}
\title{Distance-based Spatial Weights}
\usage{
st_distance_weights(
  sfj,
  unit = "km",
  dist_thres = NULL,
  power = 1,
  is_inverse = FALSE
)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{unit}{(optional) The unit for calculating spatial distance, can be
'km'(default) or 'mile'.}

\item{dist_thres}{(optional) A positive numeric value of distance threshold.}

\item{power}{(optional) The power (or exponent) of a number indicates how
many times to use the number in a multiplication.Default is 1.}

\item{is_inverse}{(optional) FALSE (default) or TRUE, apply inverse on
distance value.}
}
\value{
An instance of rgeoda Weight-class.
}
\description{
Create a distance-based weights
}
\examples{
guerry_path = system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry = sf::read_sf(guerry_path)
st_distance_weights(guerry)
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
