% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_wt.R
\name{st_kernel_knn_weights}
\alias{st_kernel_knn_weights}
\title{K-NN Kernel Spatial Weights}
\usage{
st_kernel_knn_weights(
  sfj,
  k,
  kernel = "gaussian",
  power = 1,
  adaptive_bandwidth = TRUE,
  use_kernel_diagonals = FALSE,
  is_inverse = FALSE,
  unit = "km"
)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{k}{A positive integer number for k-nearest neighbors.}

\item{kernel}{(optional) A string value, which has to be one of 'triangular',
'uniform', 'epanechnikov', 'quartic', 'gaussian'(default).}

\item{power}{(optional) The power (or exponent) of a number indicates how
many times to use the number in a multiplication.Default is 1.}

\item{adaptive_bandwidth}{(optional) TRUE (default) or FALSE: TRUE use
adaptive bandwidth calculated using distance of k-nearest neithbors,
FALSE use max distance of all observation to their k-nearest neighbors.}

\item{use_kernel_diagonals}{(optional) FALSE (default) or TRUE, apply kernel
on the diagonal of weights matrix.}

\item{is_inverse}{(optional) FALSE (default) or TRUE, apply inverse on
distance value.}

\item{unit}{(optional) The unit for calculating spatial distance, can be
'km'(default) or 'mile'.}
}
\value{
An instance of rgeoda Weight-class.
}
\description{
Create a kernel weights by specifying k-nearest neighbors and a kernel method
}
\examples{
guerry_path = system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry = sf::read_sf(guerry_path)
st_kernel_knn_weights(guerry,6)
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
