% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_wt.R
\name{st_kernel_weights}
\alias{st_kernel_weights}
\title{Distance-based Kernel Spatial Weights}
\usage{
st_kernel_weights(
  sfj,
  kernel = "gaussian",
  bandwidth = NULL,
  power = 1,
  use_kernel_diagonals = FALSE,
  is_inverse = FALSE,
  unit = "km"
)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{kernel}{(optional) A string value, which has to be one of 'triangular',
'uniform', 'epanechnikov', 'quartic', 'gaussian'(default).}

\item{bandwidth}{(optional) A positive numeric value of bandwidth.}

\item{power}{(optional) The power (or exponent) of a number indicates how
many times to use the number in a multiplication.Default is 1.}

\item{use_kernel_diagonals}{(optional) FALSE (default) or TRUE, apply kernel
on the diagonal of weights matrix.}

\item{is_inverse}{(optional) FALSE (default) or TRUE, apply inverse on
distance value.}

\item{unit}{(optional) The unit for calculating spatial distance, can be
'km'(default) or 'mile'.}
}
\value{
An instance of rgeoda Weight-class.
}
\description{
Create a kernel weights by specifying a bandwidth and a kernel method
}
\examples{
guerry_path = system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry = sf::read_sf(guerry_path)
st_kernel_weights(guerry)
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
