% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_wt.R
\name{st_knn_weights}
\alias{st_knn_weights}
\title{K-Nearest Neighbors-based Spatial Weights}
\usage{
st_knn_weights(sfj, k, power = 1, is_inverse = FALSE, unit = "km")
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{k}{A positive integer number for k-nearest neighbors.}

\item{power}{(optional) The power (or exponent) of a number indicates how
many times to use the number in a multiplication.Default is 1.}

\item{is_inverse}{(optional) FALSE (default) or TRUE, apply inverse on
distance value.}

\item{unit}{(optional) The unit for calculating spatial distance, can be
'km'(default) or 'mile'.}
}
\value{
An instance of rgeoda Weight-class.
}
\description{
Create a k-nearest neighbors based spatial weights
}
\examples{
guerry_path = system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry = sf::read_sf(guerry_path)
st_knn_weights(guerry,3)
}
\author{
Wenbo Lv
}
