% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_wt.R
\name{st_lag}
\alias{st_lag}
\title{Spatial Lag}
\usage{
st_lag(sfj, varcol, wt = NULL)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{varcol}{The variable selected to calculate spatial lag, which is a character.}

\item{wt}{(optional) The spatial weights object,which can use \code{st_weights()} to
construct,default is constructed by \code{st_weights(sfj,'contiguity')}.}
}
\value{
A numeric vector.
}
\description{
Compute the spatial lag for idx-th observation using selected variable and
spatial weights matrix
}
\examples{
guerry_path = system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry = sf::read_sf(guerry_path)
st_lag(guerry,'Pop1831')
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
