% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_wt.R
\name{st_lnmt}
\alias{st_lnmt}
\title{Local Neighbor Match Test}
\usage{
st_lnmt(
  sfj,
  varcol,
  k,
  unit = "km",
  scale_method = "standardize",
  distance_method = "euclidean",
  power = 1,
  is_inverse = FALSE
)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{varcol}{The variables selected to run local neighbor match test.}

\item{k}{A positive integer number for k-nearest neighbors searching.}

\item{unit}{(optional) The unit for calculating spatial distance, can be
'km'(default) or 'mile'.}

\item{scale_method}{(optional) One of the scaling methods 'raw', 'standardize', 'demean', 'mad',
'range_standardize', 'range_adjust' to apply on input data. Default is 'standardize' (Z-score normalization).}

\item{distance_method}{(optional) The type of distance metrics used to measure the distance between input data.
Options are 'euclidean', 'manhattan'. Default is 'euclidean'.}

\item{power}{(optional) The power (or exponent) of a number says how many times to use
the number in a multiplication.}

\item{is_inverse}{(optional) FALSE (default) or TRUE, apply inverse on distance value.}
}
\value{
A tibble with two columns "Cardinality" and "Probability".
}
\description{
The local neighbor match test is to assess the extent of overlap between k-nearest neighbors
in geographical space and k-nearest neighbors in multi-attribute space.
}
\examples{
library(sf)
guerry = read_sf(system.file("extdata","Guerry.shp",package = "rgeoda"))
st_lnmt(guerry,c('Crm_prs','Crm_prp','Litercy','Donatns',
'Infants','Suicids'),6)
}
\author{
Wenbo Lv
}
