% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_spatstat.R
\name{st_local_joincount}
\alias{st_local_joincount}
\title{Local Join Count Statistics}
\usage{
st_local_joincount(
  sfj,
  varcol,
  wt = NULL,
  permutations = 999,
  permutation_method = "complete",
  significance_cutoff = 0.05,
  cpu_threads = 6,
  seed = 123456789
)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{varcol}{The variable selected to calculate spatial lag, which is a character.}

\item{wt}{(optional) The spatial weights object,which can use \code{st_weights()} to
construct,default is constructed by \code{st_weights(sfj,'contiguity')}.}

\item{permutations}{(optional) The number of permutations for the LISA computation.}

\item{permutation_method}{(optional) The permutation method used for the LISA computation. Options are 'complete', 'lookup'.
Default is 'complete'.}

\item{significance_cutoff}{(optional) A cutoff value for significance p-values to filter
not-significant clusters.}

\item{cpu_threads}{(optional) The number of cpu threads used for parallel LISA computation.}

\item{seed}{(optional) The seed for random number generator.}
}
\value{
A factor vector.
}
\description{
The function to apply local Join Count statistics
}
\examples{
guerry_path = system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry = sf::read_sf(guerry_path)
st_local_joincount(guerry,'Crm_prp')
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
