% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_spatstat.R
\name{st_local_multiquantilelisa}
\alias{st_local_multiquantilelisa}
\title{Multivariate Quantile LISA Statistics}
\usage{
st_local_multiquantilelisa(
  sfj,
  varcol,
  k,
  q,
  wt = NULL,
  permutations = 999,
  permutation_method = "complete",
  significance_cutoff = 0.05,
  cpu_threads = 6,
  seed = 123456789
)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{varcol}{The variable selected to calculate spatial lag, which is a character.}

\item{k}{A value indicates the number of quantiles. Value range e.g. \verb{[1, 10]}.}

\item{q}{A value indicates which quantile or interval used in local join count statistics.
Value stars from 1.}

\item{wt}{(optional) The spatial weights object,which can use \code{st_weights()} to
construct,default is constructed by \code{st_weights(sfj,'contiguity')}.}

\item{permutations}{(optional) The number of permutations for the LISA computation.}

\item{permutation_method}{(optional) The permutation method used for the LISA computation. Options are 'complete', 'lookup'.
Default is 'complete'.}

\item{significance_cutoff}{(optional) A cutoff value for significance p-values to filter
not-significant clusters.}

\item{cpu_threads}{(optional) The number of cpu threads used for parallel LISA computation.}

\item{seed}{(optional) The seed for random number generator.}
}
\value{
A factor vector.
}
\description{
Function to apply multivariate quantile LISA statistics
}
\examples{
guerry_path = system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry = sf::read_sf(guerry_path)
st_local_multiquantilelisa(guerry,c("Crm_prp", "Litercy"),c(4,4), c(1,1))
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
