% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{st_maxp_greedy}
\alias{st_maxp_greedy}
\title{A greedy algorithm to solve the max-p-region problem}
\usage{
st_maxp_greedy(
  sfj,
  varcol,
  wt = NULL,
  boundvar,
  min_bound,
  iterations = 99,
  initial_regions = vector("numeric"),
  scale_method = "standardize",
  distance_method = "euclidean",
  seed = 123456789,
  cpu_threads = 6,
  rdist = numeric()
)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{varcol}{The variable selected to calculate spatial lag, which is a character.}

\item{wt}{(optional) The spatial weights object,which can use \code{st_weights()} to
construct,default is constructed by \code{st_weights(sfj,'contiguity')}.}

\item{boundvar}{A numeric vector of selected bounding variable.}

\item{min_bound}{A minimum value that the sum value of bounding variable int each
cluster should be greater than.}

\item{iterations}{(optional) The number of iterations of greedy algorithm. Defaults to 99.}

\item{initial_regions}{(optional) The initial regions that the local search starts with.
Default is empty. means the local search starts with a random process to "grow" clusters.}

\item{scale_method}{(optional) One of the scaling methods 'raw', 'standardize',
'demean', 'mad', 'range_standardize', 'range_adjust' to apply on input data.
Default is 'standardize' (Z-score normalization).}

\item{distance_method}{(optional) The distance method used to compute the distance betwen
observation i and j. Defaults to "euclidean". Options are "euclidean" and "manhattan"}

\item{seed}{(optional) The seed for random number generator. Defaults to 123456789.}

\item{cpu_threads}{(optional) The number of cpu threads used for parallel computation.Default is 6.}

\item{rdist}{(optional) The distance matrix (lower triangular matrix, column wise storage).}
}
\value{
A names list with names "Clusters", "Total sum of squares", "Within-cluster sum of squares",
"Total within-cluster sum of squares", and "The ratio of between to total sum of squares".
}
\description{
A wrapper function for \code{rgeoda::maxp_greedy()}.The max-p-region problem is a special case
of constrained clustering where a finite number of geographical areas are aggregated into
the maximum number of regions (max-p-regions), such that each region is geographically
connected and the clusters could maximize internal homogeneity.
}
\examples{
library(sf)
guerry = read_sf(system.file("extdata", "Guerry.shp", package = "rgeoda"))
guerry_clusters = st_maxp_greedy(guerry,c('Crm_prs','Crm_prp','Litercy','Donatns',
'Infants','Suicids'),boundvar = 'Pop1831',min_bound = 3236.67)
guerry_clusters
}
\author{
Wenbo Lv
}
