% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{st_schc}
\alias{st_schc}
\title{Spatially Constrained Hierarchical Clucstering (SCHC)}
\usage{
st_schc(
  sfj,
  varcol,
  k,
  wt = NULL,
  boundvar = NULL,
  method = "average",
  min_bound = 0,
  scale_method = "standardize",
  distance_method = "euclidean",
  rdist = numeric()
)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{varcol}{The variable selected to calculate spatial lag, which is a character.}

\item{k}{The number of clusters.}

\item{wt}{(optional) The spatial weights object,which can use \code{st_weights()} to
construct,default is constructed by \code{st_weights(sfj,'contiguity')}.}

\item{boundvar}{(optional) A data frame / tibble with selected bound variable.}

\item{method}{(optional) "single", "complete", "average"(default),"ward".}

\item{min_bound}{(optional) A minimum bound value that applies to all clusters.}

\item{scale_method}{(optional) One of the scaling methods 'raw', 'standardize',
'demean', 'mad', 'range_standardize', 'range_adjust' to apply on input data.
Default is 'standardize' (Z-score normalization).}

\item{distance_method}{(optional) The distance method used to compute the distance
between observation i and j. Defaults to "euclidean". Options are "euclidean" and "manhattan".}

\item{rdist}{(optional) The distance matrix (lower triangular matrix, column wise storage).}
}
\value{
A names list with names "Clusters", "Total sum of squares", "Within-cluster sum of squares",
"Total within-cluster sum of squares", and "The ratio of between to total sum of squares".
}
\description{
A wrapper function for \code{rgeoda::schc()}.Spatially constrained hierarchical clustering is
a special form of constrained clustering, where the constraint is based on contiguity (common borders).
The method builds up the clusters using agglomerative hierarchical clustering methods: single linkage,
complete linkage, average linkage and Ward's method (a special form of centroid linkage). Meanwhile,
it also maintains the spatial contiguity when merging two clusters.
}
\examples{
library(sf)
guerry = read_sf(system.file("extdata", "Guerry.shp", package = "rgeoda"))
guerry_clusters = st_schc(guerry,c('Crm_prs','Crm_prp','Litercy','Donatns','Infants','Suicids'),
4,method = "complete")
guerry_clusters
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
