% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{st_skater}
\alias{st_skater}
\title{Spatial C(K)luster Analysis by Tree Edge Removal(SKATER)}
\usage{
st_skater(
  sfj,
  varcol,
  k,
  wt = NULL,
  boundvar = NULL,
  min_bound = 0,
  scale_method = "standardize",
  distance_method = "euclidean",
  seed = 123456789,
  cpu_threads = 6,
  rdist = numeric()
)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{varcol}{The variable selected to calculate spatial lag, which is a character.}

\item{k}{The number of clusters.}

\item{wt}{(optional) The spatial weights object,which can use \code{st_weights()} to
construct,default is constructed by \code{st_weights(sfj,'contiguity')}.}

\item{boundvar}{(optional) A data frame / tibble with selected bound variable.}

\item{min_bound}{(optional) A minimum bound value that applies to all clusters.}

\item{scale_method}{(optional) One of the scaling methods 'raw', 'standardize',
'demean', 'mad', 'range_standardize', 'range_adjust' to apply on input data.
Default is 'standardize' (Z-score normalization).}

\item{distance_method}{(optional) The distance method used to compute the distance
between observation i and j. Defaults to "euclidean". Options are "euclidean" and "manhattan".}

\item{seed}{(int,optional) The seed for random number generator. Defaults to 123456789.}

\item{cpu_threads}{(optional) The number of cpu threads used for parallel computation.}

\item{rdist}{(optional) The distance matrix (lower triangular matrix, column wise storage).}
}
\value{
A names list with names "Clusters", "Total sum of squares", "Within-cluster sum of squares",
"Total within-cluster sum of squares", and "The ratio of between to total sum of squares".
}
\description{
A wrapper function for \code{rgeoda::skater()}.SKATER forms clusters by spatially
partitioning data that has similar values for features of interest.
}
\examples{
library(sf)
guerry = read_sf(system.file("extdata", "Guerry.shp", package = "rgeoda"))
guerry_clusters = st_skater(guerry,c('Crm_prs','Crm_prp','Litercy','Donatns','Infants','Suicids'),4)
guerry_clusters
}
\author{
Wenbo Lv
}
