% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_wt.R
\name{st_summary}
\alias{st_summary}
\title{Summary of Spatial Weights}
\usage{
st_summary(wt, ...)
}
\arguments{
\item{wt}{A Weight object}

\item{...}{summary optional parameters}
}
\value{
A summary description of an instance of Weight-class
}
\description{
Warpping the \code{summary()} function for spatial weights
}
\examples{
\dontrun{
library(sf)
guerry_path = system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry = read_sf(guerry_path)
queen_w = tidyrgeoda::st_weights(guerry,'contiguity')
st_summary(queen_w)
}
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
