% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_wt.R
\name{st_weights}
\alias{st_weights}
\title{Construct Spatial Weights}
\usage{
st_weights(sfj, weight = NULL, ...)
}
\arguments{
\item{sfj}{An sf (simple feature) object.}

\item{weight}{The method used to create spatial weights,which has to be one
of 'contiguity', 'distance', 'knn', 'kernel', 'kernel_knn'.}

\item{...}{Other arguments to construct spatial weight, see
'tidyrgeoda::st_contiguity_weights','tidyrgeoda::st_distance_weights',
'tidyrgeoda::st_knn_weights','tidyrgeoda::st_kernel_weights',
'tidyrgeoda::st_kernel_knn_weights'.}
}
\value{
An instance of rgeoda Weight-class.
}
\description{
Create a spatial weights
}
\examples{
guerry_path = system.file("extdata", "Guerry.shp", package = "rgeoda")
guerry = sf::read_sf(guerry_path)
st_weights(guerry,'kernel_knn',6)
}
\author{
Wenbo Lv
}
