% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_wt.R
\name{write_geoda}
\alias{write_geoda}
\title{Save Spatial Weights}
\usage{
write_geoda(wt, dsn, layer = NULL, id_vec = NULL)
}
\arguments{
\item{wt}{A Weight object}

\item{dsn}{The path of an output weights file}

\item{layer}{(optional) The name of the layer of input dataset,default is \code{""}.}

\item{id_vec}{(optional) Defines the unique value of each observation when saving a
weights file. Default is \code{tibble::tibble(id_v = 1:wt$num_obs)}.}
}
\value{
A boolean value indicates if save successfully or failed
}
\description{
Save spatial weights to a file
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
